+++ smart-contracts/contracts-common/concordium-contracts-common/src/types.rs
@@ -46,6 +46,8 @@ pub type ContractSubIndex = u64;
 /// NB: This is different from the Base58 representation.
 pub const ACCOUNT_ADDRESS_SIZE: usize = 32;
 
+const CANONICAL_ACCOUNT_ADDRESS_SIZE: usize = 29;
+
 /// The type of amounts on the chain.
 #[repr(transparent)]
 #[derive(Copy, Clone, Debug, Default, PartialEq, Eq, PartialOrd, Ord, Hash)]
@@ -1062,6 +1064,14 @@ impl<'de> SerdeDeserialize<'de> for Duration {
     }
 }
 
+/// Canonical address of an account, as raw bytes.
+/// The canonical address is the first 29 bytes of the account address, uniquely
+/// identifying accounts. The last 3 bytes is reserved as an account alias, to
+/// be used for example by exchanges to uniquely identify graceful clients
+#[derive(Eq, PartialEq, Copy, Clone, PartialOrd, Ord, Debug, Hash)]
+#[repr(transparent)]
+pub struct CanonicalAccountAddress(pub [u8; CANONICAL_ACCOUNT_ADDRESS_SIZE]);
+
 /// Address of an account, as raw bytes.
 #[derive(Eq, PartialEq, Copy, Clone, PartialOrd, Ord, Debug, Hash)]
 #[cfg_attr(feature = "fuzz", derive(Arbitrary))]
@@ -1087,10 +1097,23 @@ impl convert::AsMut<[u8; 32]> for AccountAddress {
 }
 
 impl AccountAddress {
+    /// Get the canonical address representing the unique first 29 bytes of the
+    /// account address. This is the unique account address part and is
+    /// independent of the individual aliases.
+    pub fn get_canonical_address(&self) -> CanonicalAccountAddress {
+        CanonicalAccountAddress(
+            self.0[..CANONICAL_ACCOUNT_ADDRESS_SIZE]
+                .try_into()
+                .expect("Slice with incorrect length"),
+        )
+    }
+
     /// Check whether `self` is an alias of `other`. Two addresses are aliases
     /// if they identify the same account. This is defined to be when the
     /// addresses agree on the first 29 bytes.
-    pub fn is_alias(&self, other: &AccountAddress) -> bool { self.0[0..29] == other.0[0..29] }
+    pub fn is_alias(&self, other: &AccountAddress) -> bool {
+        self.0[..CANONICAL_ACCOUNT_ADDRESS_SIZE] == other.0[..CANONICAL_ACCOUNT_ADDRESS_SIZE]
+    }
 
     /// Get the `n-th` alias of an address. There are 2^24 possible aliases.
     /// If the counter is `>= 2^24` then this function will return [`None`].